+++ egs2/ljspeech/tts1/conf/tuning/train_jets.yaml
@@ -74,7 +74,7 @@ tts_conf:
         generator_nonlinear_activation_params:
             negative_slope: 0.1
         generator_use_weight_norm: true
-        segment_size: 64                             # segment size for random windowed discriminator
+        segment_size: 32                             # segment size for random windowed discriminator
 
     # discriminator related
     discriminator_type: hifigan_multi_scale_multi_period_discriminator
@@ -140,7 +140,7 @@ tts_conf:
     lambda_mel: 45.0                     # loss scaling coefficient for Mel loss
     lambda_feat_match: 2.0               # loss scaling coefficient for feat match loss
     lambda_var: 1.0
-    lambda_align: 2.0
+    lambda_align: 1.0
     # others
     sampling_rate: 22050                 # needed in the inference for saving wav
     cache_generator_outputs: true        # whether to cache generator outputs in the training
